Move error handlers to their own package
